Updated saving of EXRs to open a requester with options. Made data window square appear in stencil region, too. Made image width and height be ruled by the dataWindow first.
Former-commit-id: 75444b4a521fd47bd16c8427a4e57f410ff5a5c5

+++ mrViewer/src/video/mrvGLShape.h
@@ -21,7 +21,7 @@ class Point
      {
      }
 
-     Point( int xx, int yy ) :
+     Point( double xx, double yy ) :
      x( xx ), y( yy )
      {
      }
